imp(finder): organize structure, add base class provider
