+++ src/js/finder.js
@@ -1,62 +1,9 @@
-import Modal from "./modal.js";
+import FinderBase from "./finder-base.js";
+import { flatMeshToSingleArray } from "./utils.js";
 
-class Finder {
+class Finder extends FinderBase {
   constructor(meshSize) {
-    this.meshSize = meshSize;
-    this.startAndEndCords = {};
-    this.generateHTMLElements();
-    this.generateMesh();
-    this.generateSteps();
-    this.setStep(this.finderStep.begin);
-    this.initFieldClick();
-    this.initBtn();
-    this.initRandomRoutesButton();
-  }
-
-  generateSteps() {
-    this.finderStep = {
-      begin: {
-        step: "begin",
-        btnText: "Finish drawing",
-        headerText: "DRAW ROUTES",
-      },
-      compute: {
-        step: "compute",
-        btnText: "Compute",
-        headerText: "PICK START AND FINISH",
-      },
-      replay: {
-        step: "replay",
-        btnText: "Start again",
-        headerText: "THE BEST ROUTE IS...",
-      },
-    };
-  }
-
-  generateHTMLElements() {
-    this.drawingContainer = document.querySelector(".drawing-container");
-    this.stepTitle = document.querySelector(".step-title");
-    this.btnStep = document.querySelector(".btnChangeStep");
-    this.btnRandomRoutes = document.querySelector(".btnRandomRoutes");
-  }
-
-  generateMesh() {
-    const { drawingContainer, meshSize } = this;
-    this.mesh = {};
-
-    for (let i = 1; i <= meshSize; i++) {
-      this.mesh[i] = {};
-      for (let j = 1; j <= meshSize; j++) {
-        this.mesh[i][j] = false;
-        drawingContainer.innerHTML += `<div class="field" data-row="${i}" data-column="${j}"></div>`;
-      }
-    }
-  }
-
-  flatMeshToSingleArray(mesh) {
-    return Object.values(mesh)
-      .map((e) => Object.values(e))
-      .flat();
+    super(meshSize);
   }
 
   toggleField(clickedField) {
@@ -88,7 +35,7 @@ class Finder {
       mesh[field.row][field.column] = false;
       clickedField.classList.remove("selected");
     } else {
-      if (this.flatMeshToSingleArray(mesh).includes(true)) {
+      if (flatMeshToSingleArray(mesh).includes(true)) {
         if (!edgeFields.includes(true)) {
           alert(
             "A new field should touch at least one that is already selected!"
@@ -99,6 +46,12 @@ class Finder {
       mesh[field.row][field.column] = true;
       clickedField.classList.add("selected");
     }
+  }
+
+  selectFieldElem(row, column) {
+    return document.querySelector(
+      `[data-row="${row}"][data-column="${column}"]`
+    );
   }
 
   choseStartAndFinish(clickedField) {
@@ -167,21 +120,8 @@ class Finder {
     }
   }
 
-  initFieldClick() {
-    this.drawingContainer.addEventListener("click", (e) => {
-      const { currentStep, finderStep } = this;
-      const clickedField = e.target;
-
-      if (!clickedField.classList.contains("field")) return;
-
-      if (currentStep === finderStep.begin.step) this.toggleField(clickedField);
-      if (currentStep === finderStep.compute.step)
-        this.choseStartAndFinish(clickedField);
-    });
-  }
-
   computeBestRoute() {
-    const { startAndEndCords, mesh, meshSize } = this;
+    const { startAndEndCords, mesh, meshSize, directions } = this;
     const { start: startCords, end: endCords } = startAndEndCords;
 
     const result = [];
@@ -201,13 +141,6 @@ class Finder {
         result.push(path);
         continue;
       }
-
-      const directions = [
-        [-1, 0],
-        [1, 0],
-        [0, -1],
-        [0, 1],
-      ];
 
       for (const [x, y] of directions) {
         const r = row + x;
@@ -238,7 +171,7 @@ class Finder {
         a.length < b.length ? a : b
       );
 
-      this.fullMeshLength = this.flatMeshToSingleArray(mesh).filter(
+      this.fullMeshLength = flatMeshToSingleArray(mesh).filter(
         (field) => field === true
       ).length;
 
@@ -260,126 +193,10 @@ class Finder {
     }
   }
 
-  initBtn() {
-    this.btnStep.addEventListener("click", () => {
-      const {
-        finderStep,
-        currentStep,
-        startAndEndCords,
-        mesh,
-        flatMeshToSingleArray,
-      } = this;
-
-      if (currentStep === finderStep.begin.step) {
-        const flatMesh = flatMeshToSingleArray(mesh);
-        if (flatMesh.filter((field) => field === true).length < 2) {
-          alert("Pleas select at least 2 fields");
-          return;
-        }
-        this.setStep(finderStep.compute);
-        return;
-      }
-      if (currentStep === finderStep.compute.step) {
-        if (!("start" in startAndEndCords) || !("end" in startAndEndCords)) {
-          alert("Pls chose a start and finish for route");
-        } else {
-          const computed = this.computeBestRoute();
-          this.setStep(finderStep.replay);
-
-          if (computed) {
-            document.body.scrollTo(0, 0);
-            document.body.style.overflow = "hidden";
-
-            new Modal([
-              this.fullMeshLength,
-              this.shortestRouteLength,
-              this.longestRouteLength,
-            ]);
-          }
-
-          return;
-        }
-      }
-      if (currentStep === finderStep.replay.step) {
-        this.resetFinder();
-        this.setStep(finderStep.begin);
-        return;
-      }
-    });
-  }
-
-  initRandomRoutesButton() {
-    this.btnRandomRoutes.addEventListener("click", () => {
-      this.generateRandomRoutes();
-    });
-  }
-
   resetFinder() {
     this.startAndEndCords = {};
     this.drawingContainer.innerHTML = "";
     this.generateMesh();
-  }
-
-  setStep({ step, headerText, btnText }) {
-    this.stepTitle.textContent = headerText;
-    this.btnStep.textContent = btnText;
-    this.currentStep = step;
-  }
-
-  selectFieldElem(row, column) {
-    return document.querySelector(
-      `[data-row="${row}"][data-column="${column}"]`
-    );
-  }
-
-  generateRandomRoutes() {
-    const { mesh, meshSize, finderStep } = this;
-
-    this.drawingContainer
-      .querySelectorAll(".field.selected, .field.success, .field.longest")
-      .forEach((field) => {
-        field.classList.remove("selected", "success", "longest");
-        field.textContent = "";
-      });
-
-    for (let i = 1; i <= meshSize; i++) {
-      for (let j = 1; j <= meshSize; j++) {
-        mesh[i][j] = false;
-      }
-    }
-
-    this.setStep(finderStep.begin);
-
-    for (let i = 0; i < meshSize * 5; i++) {
-      let randomRow, randomColumn;
-
-      do {
-        randomRow = Math.floor(Math.random() * meshSize) + 1;
-        randomColumn = Math.floor(Math.random() * meshSize) + 1;
-      } while (!this.isValidRandomRoute(randomRow, randomColumn));
-
-      mesh[randomRow][randomColumn] = true;
-      this.selectFieldElem(randomRow, randomColumn).classList.add("selected");
-    }
-  }
-
-  isValidRandomRoute(row, column) {
-    const { mesh, flatMeshToSingleArray } = this;
-
-    if (flatMeshToSingleArray(mesh).includes(true)) {
-      const edgeFields = [
-        mesh[row - 1]?.[column],
-        mesh[row + 1]?.[column],
-        mesh[row]?.[column - 1],
-        mesh[row]?.[column + 1],
-      ];
-
-      if (!edgeFields.includes(true)) {
-        return false;
-      }
-    }
-
-    return true;
   }
 }
 export default Finder;
