+++ src/js/utils.js
@@ -1,0 +1,5 @@
+export function flatMeshToSingleArray(mesh) {
+  return Object.values(mesh)
+    .map((e) => Object.values(e))
+    .flat();
+}
